feat(date-logic): default weekday is now short

+++ packages/date-logic/index.js
@@ -3,7 +3,12 @@
  */
 module.exports = function getDate(
   locale = "en-US",
-  options = { weekday: "long", year: "numeric", month: "short", day: "numeric" }
+  options = {
+    weekday: "short",
+    year: "numeric",
+    month: "short",
+    day: "numeric"
+  }
 ) {
   return new Date().toLocaleDateString(locale, options);
 };
